Fixed submission form. Working with file attachments

+++ src/uk/ac/ebi/age/admin/server/mng/UploadManager.java
@@ -37,7 +37,10 @@ public class UploadManager
   finally
   {
    for(File f : upReq.getFiles().values())
-    f.delete();
+   {
+    if( f.exists() )
+     f.delete();
+   }
   }
  }
